ポスタープレイヤー同期

+++ Paperona4ネット/Papera4ネット/Paper/client/source/Net/Socket.h
@@ -38,8 +38,8 @@ struct PlayerData
 
 	Vector3 pos;
 	float angleY;
-	int motion_no;
-	int action_part;
+	BYTE motion_no;
+	BYTE action_part;
 };
 //---------------------------------------------------------------------
 //  プレイヤーの返ってくるパラメータ
